Change API for the maze module.
Add a new dictionary containing the list of generating algorithm.
Add four boolean variables to a Cell representing the state of the walls.
Add a variable to the Maze object which contains the cell involved in the last
operation of the maze generation.
The test script have been modified accordingly.